+++ packages/crdt_socket_sync/lib/src/client/client.dart
@@ -62,8 +62,7 @@ abstract class CRDTSocketClient {
   Future<void> sendChange(Change change);
 
   /// Request a snapshot from the server
-  // TODO(mattia): request server state like the handshake
-  Future<void> requestSnapshot();
+  Future<void> requestSync();
 
   /// Dispose the client
   void dispose();