+++ packages/crdt_socket_sync/lib/src/server/client_session_event.dart
@@ -17,6 +17,9 @@ enum SessionEventType {
   /// Client disconnected
   disconnected,
 
+  /// Client is out of sync
+  clientOutOfSync,
+
   /// Error occurred
   error,
 }