+++ packages/crdt_socket_sync/CHANGELOG.md
@@ -3,10 +3,12 @@
 **Breaking changes**
 - `CRDTServerRegistry.addDocument` now only takes a `documentId` parameter
 - `CRDTServerRegistry` methods now return a `Future`
+- rename client `requestSnapshot` to `requestSync`
 
 ### Added
 - Feature: add `messageCodec` parameter to `WebSocketServer` and `WebSocketClient`
 - Feature: `JsonMessageCodec` now supports `toEncodable` and `reviver` parameters
+- Feature: added out of sync error handling
 
 ## [0.2.0] - 2025-06-26